+++ R/dplyr_verbs.R
@@ -9,18 +9,23 @@
 #' * `*_join()`: joins an external table to the *segments* table of the eeg_lst.
 #' * `mutate()` and `transmute()` Variables that are a function of a channel are added to the signal_tbl table, and other variables are added to the segments table.
 #' * `summarize()` summarizes the channel of the signal_tbl table.
+#' * `pull()` only pulls columns of the signal table
 #'
 #' In addition, `_at()`, and `_if()` versions of the functions should work as well. Notice that  the `_at()` versions are 
 #' much faster than the `_if()` versions of these commands.
+#' 
+#' Notice that these verbs have different defaults for `eeg_lst` 
 #'
 #' @param .data An eeg_lst.
 #' @param x An eeg_lst.
 #' @param y A data frame, tibble, or data.table.
 #' @inheritParams dplyr::join
+#' @inheritParams dplyr::pull
 #' @param ... Name-value pairs of expressions; see [dplyr][dplyr::dplyr] for more help.
 #' @param .preserve Not in use, for compatibility reasons.
-#' @param add Not in use, for compatibility reasons.
-#' @param .drop When .drop = TRUE, empty groups are dropped. (FALSE by default.)
+#' @param .add When FALSE, the default, group_by() will override existing groups. To add to the existing groups, use .add = TRUE.
+#' @param .drop Only .drop = FALSE is avaialble, empty groups are never dropped.
+#' @param .groups Only .groups = "keep" is avaialble.  Same grouping structure as .data.
 #' @importFrom dplyr  select mutate transmute summarise rename
 #' @importFrom dplyr group_by ungroup group_vars
 #' @importFrom dplyr groups
@@ -77,15 +82,16 @@ NULL
 #' @rdname dplyr_verbs
 #' @export
 mutate.eeg_lst <- function(.data, ...) {
-  dots <- rlang::quos(...)
-  mutate_eeg_lst(.data, dots, keep_cols = TRUE)
+  # dots <- rlang::quos(...)
+  mutate_eeg_lst(.data, ..., keep_cols = TRUE) %>% 
+    validate_eeg_lst()
 }
 
 #' @rdname dplyr_verbs
 #' @export
 transmute.eeg_lst <- function(.data, ...) {
-  dots <- rlang::quos(...)
-  mutate_eeg_lst(.data, dots, keep_cols = FALSE)
+  # dots <- rlang::quos(...)
+  mutate_eeg_lst(.data, ..., keep_cols = FALSE)
 }
 #' @rdname dplyr_verbs
 filter.eeg_lst <- function(.data, ..., .preserve = FALSE) {
@@ -104,18 +110,21 @@ filter.eeg_ica_lst <- function(.data, ..., .preserve = FALSE) {
 }
 #' @rdname dplyr_verbs
 #' @export
-summarise.eeg_lst <- function(.data, ...) {
+summarise.eeg_lst <- function(.data, ..., .groups = "keep") {
   dots <- rlang::quos(...)
-  summarize_eeg_lst(.data, dots)
+  if(.groups != "keep") {
+    warning("Only  'keep' option is available")
+  }
+  summarize_eeg_lst(.data, dots, .groups = "keep")
 }
 #' @rdname dplyr_verbs
 #' @export
-group_by.eeg_lst <- function(.data, ..., add = FALSE, .drop = FALSE) {
+group_by.eeg_lst <- function(.data, ..., .add = FALSE, .drop = FALSE) {
   dots <- rlang::quos(...)
   if (.drop == TRUE) {
-    warning("Ignoring .drop argument. It hasn't been implemented yet.")
+    warning("Ignoring .drop argument. It can only be set to FALSE.")
   }
-  group_by_eeg_lst(.eeg_lst = .data, dots, .add = add)
+  group_by_eeg_lst(.eeg_lst = .data, dots, .add = .add)
 }
 #' @rdname dplyr_verbs
 #' @export
@@ -183,8 +192,21 @@ semi_join.eeg_lst <- function(x, y, by = NULL, suffix = c(".x", ".y"), ...) {
   x$.events <- semi_join_dt(x$.events, segments, by = ".id")
   x %>% validate_eeg_lst()
 }
+
 #' @rdname dplyr_verbs
 #' @export
 tbl_vars.eeg_lst <- function(x) {
   setdiff(dplyr::tbl_vars(x$.signal), c(dplyr::tbl_vars(x$.segments), c(".id", ".sample")))
+}
+
+#' @rdname dplyr_verbs
+#' @export
+pull.eeg_lst <- function(.data, var = -1, name = NULL, ...) {
+  var <- tidyselect::vars_pull(names(.data$.signal), !!rlang::enquo(var))
+  name <- rlang::enquo(name)
+  if (rlang::quo_is_null(name)) {
+    return(.data$.signal[[var]])
+  }
+  name <- tidyselect::vars_pull(names(.data$.signal), !!name)
+  rlang::set_names(.data[[var]], nm = .data[[name]])
 }