+++ R/filters.R
@@ -1,6 +1,14 @@
 #' Apply a zero-phase low-pass, high-pass, band-pass, or band-stop FIR filter.
 #'
-#' Apply a zero-phase low-pass, high-pass, band-pass, or band-stop filter of the  FIR (finite impulse response) class to every segment of an `eeg_lst`. These filters are adapted from the default filters in [MNE package](https://mne-tools.github.io) (v 0.0.17.1)  of [python](https://www.python.org/). For background information about the FIR vs IIR filters, see [here](https://martinos.org/mne/dev/auto_tutorials/plot_background_filtering.html#sphx-glr-auto-tutorials-plot-background-filtering-py).
+#' Apply a zero-phase low-pass, high-pass, band-pass, or band-stop filter of the  
+#' FIR (finite impulse response) class to every segment of an `eeg_lst`. 
+#' **It's advisable to avoid filtering short segments of data.** In general, these filters are relatively fast, a
+#'  considerable slowdown can be due to incorrectly specified filters (e.g., confusing
+#'   low with high pass filters) or to filtering too many short segments. All these filters
+#'  are adapted from the default filters in
+#'   [MNE package](https://mne.tools/0.17/generated/mne.filter.filter_data.html#mne.filter.filter_data) (v 0.0.17.1)
+#'     of [python](https://www.python.org/). For background information
+#'      about the FIR vs IIR filters, see [here](https://mne.tools/dev/auto_tutorials/discussions/plot_background_filtering.html).
 #'
 #' * `eeg_filt_low_pass()` Low-pass or high-cut filter.
 #' * `eeg_filt_high_pass()` High-pass or low-cut filter.
@@ -12,7 +20,10 @@
 #' @param freq A single cut frequency for `eeg_filt_low_pass` and `eeg_filt_high_pass`, two edges for
 #'   `eeg_filt_band_pass` and `eeg_filt_band_stop`.
 #' @param ... Channels to apply the filters to. All the channels by default.
-#' @param config Other parameters passed in a list to the ICA method. (Not implemented)
+#' @param config Other parameters passed in a list to the method. At the moment, only the 
+#' width of the transition band at the (high and/or low) cut-off frequencies can
+#'  be modified: `l_trans_bandwidth`, `h_trans_bandwidth`, by default they are set to "auto". 
+#'  For now the filters can only have a phase of zero (the delay for the filter is compensated).
 #' @param na.rm =TRUE will set to NA the entire segment that contains an NA, otherwise the filter will stop with an error.
 #' @return A channel or an eeg_lst.
 #' @family preprocessing functions
@@ -22,7 +33,7 @@
 #' library(ggplot2)
 #' data("data_faces_ERPs")
 #' data_ERPs_filtered <- data_faces_ERPs %>%
-#'   eeg_filt_low_pass(freq = 1)
+#'   eeg_filt_low_pass(freq = 50)
 #' # Compare the ERPs
 #' data_faces_ERPs %>%
 #'   select(O1, O2, P7, P8) %>%