+++ tests/testthat/test_reading.R
@@ -63,7 +63,7 @@ channels_tbl(edf_f) <- ch_tbl
 channels_tbl(edf) <- channels_tbl(edf_bv)
 channels_tbl(edf_plus_bv) <- channels_tbl(edf_bv)
 events_bv <- events_tbl(multiplexed_bin_bv2) %>%
-  dplyr::mutate(.type = NA, .description = ifelse(.description == "", "New Segment", .description)) %>%
+  dplyr::mutate(.type = NA_character_, .description = ifelse(.description == "", "New Segment", .description)) %>%
   dplyr::as_tibble()
 events_edf <- events_tbl(edf_plus_bv) %>% dplyr::as_tibble()
 
@@ -86,10 +86,10 @@ test_that("edf and bdf files match", {
 
 test_that("bdf  match the bdf created by MNE", {
   bdf_file <- system.file("testdata", "Newtest17-256.bdf", package = "eeguana")
-  mne_bdf <- eeguana:::mne_bdf
+  data_mne_bdf <- eeguana:::data_mne_bdf
   bdf <- read_edf(file = bdf_file, .recording = "bdf")
 
-  expect_equal(bdf, mne_bdf)
+  expect_equal(bdf, data_mne_bdf)
 })
 
 ## segmented
@@ -110,3 +110,4 @@ test_that("seg matches", {
   expect_equal(events_tbl(seg_bin_bv2)[.type == "Stimulus"], events_tbl(seged_bin))
   expect_equal(seg_bin_bv2$.segments, dplyr::select(seged_bin$.segments, -type, -description))
 })
+