+++ tests/testthat/test_tidyverse_mutate.R
@@ -2,39 +2,7 @@ context("test tidyverse dplyr::mutate")
 library(eeguana)
 
 # tests when factors are used should be done.
-
-data_1 <- eeg_lst(
-  signal_tbl =
-    dplyr::tibble(
-      X = sin(1:30), Y = cos(1:30),
-      .id = rep(c(1L, 2L, 3L), each = 10),
-      .sample = sample_int(rep(seq(-4L, 5L), times = 3), sampling_rate = 500)
-    ),
-  channels_tbl = dplyr::tibble(
-    .channel = c("X", "Y"), .reference = NA, theta = NA, phi = NA,
-    radius = NA, .x = c(1, 1), .y = NA_real_, .z = NA_real_
-  ),
-  events_tbl = dplyr::tribble(
-    ~.id, ~.type, ~.description, ~.initial, ~.final, ~.channel,
-    1L, "New Segment", NA_character_, -4L, -4L, NA,
-    1L, "Bad", NA_character_, -2L, 0L, NA,
-    1L, "Time 0", NA_character_, 1L, 1L, NA,
-    1L, "Bad", NA_character_, 2L, 3L, "X",
-    2L, "New Segment", NA_character_, -4L, -4L, NA,
-    2L, "Time 0", NA_character_, 1L, 1L, NA,
-    2L, "Bad", NA_character_, 2L, 2L, "Y",
-    3L, "New Segment", NA_character_, -4L, -4L, NA,
-    3L, "Time 0", NA_character_, 1L, 1L, NA,
-    3L, "Bad", NA_character_, 2L, 2L, "Y"
-  ),
-  segments_tbl = dplyr::tibble(
-    .id = c(1L, 2L, 3L),
-    .recording = "recording1",
-    segment = c(1L, 2L, 3L),
-    condition = c("a", "b", "a")
-  )
-)
-
+data_1 <- eeguana:::data_sincos3id
 
 # just some different X and Y
 data_2 <- dplyr::mutate(data_1,
@@ -94,6 +62,16 @@ transmute_tbl <- data %>%
   dplyr::filter(.key == "X") %>%
   dplyr::transmute(X = .value + 1)
 
+nsamples <- 100
+nsamples_ <- 100
+data_nsamples_ <- data %>% dplyr::mutate(Z = X + nsamples_)
+data_100 <- data %>% dplyr::mutate(Z = X + 100)
+data_nsamples <- data %>% dplyr::mutate(Z = X + nsamples)
+
+test_that("dplyr:mutate functions understand the right scope",{
+  expect_equal(data_100, data_nsamples)
+  expect_equal(data_nsamples_, data_nsamples)
+})
 
 test_that("dplyr::mutate functions work correctly on ungrouped data", {
   expect_equal(as.double(mutate_eeg_lst$.signal[["X"]]), mutate_tbl$X)
@@ -150,6 +128,7 @@ group4_by_eeg_lst <- dplyr::group_by(data, .sample, .recording)
 group5_by_eeg_lst <- dplyr::group_by(data, .id, .recording)
 group6_by_eeg_lst <- dplyr::group_by(data, .id, .sample, .recording)
 group7_by_eeg_lst <- dplyr::group_by(data, .sample, condition)
+group8_by_eeg_lst <- dplyr::mutate(data, trial = segment) %>% dplyr::group_by(trial)
 
 
 mutate_g_signal_eeg <- dplyr::mutate(group_by_eeg_lst, X = X + 1)
@@ -296,6 +275,11 @@ test_that("the classes of channels of signal_tbl remain in grouped eeg_lst", {
   expect_equal(is_channel_dbl(transmute_g_signal_eeg$.signal$X), TRUE)
   expect_equal(is_channel_dbl(mutate_all_g_signal_eeg$.signal$X), TRUE)
   expect_equal(is_channel_dbl(mutate_at_g_signal_eeg$.signal$X), TRUE)
+})
+
+test_that("use a group in mutate",{
+expect_equal(group8_by_eeg_lst %>% dplyr::mutate(X = X * trial) %>% dplyr::ungroup(),
+  data %>% dplyr::mutate(trial= segment) %>% dplyr::mutate(X = X * trial))
 })